+++ mkdocs_pdf_generate/templates/filters/datetime.py
@@ -1,9 +1,0 @@
-from datetime import date, datetime
-
-
-def strptime(value: str, fmt) -> date:
-    return datetime.strptime(value, fmt)
-
-
-def strftime(value, fmt) -> str:
-    return value.strftime(fmt)