+++ mkdocs_pdf_generate/templates/template.py
@@ -1,17 +1,21 @@
 # import os
 from datetime import datetime
+from typing import Dict, List, Any, Optional, MutableMapping
 from pathlib import Path
 
 import jinja2
-from mkdocs.config.base import Config
+from mkdocs.config.defaults import MkDocsConfig
 
-from .filters.datetime import strftime, strptime
+from .filters.datetime_filter import strftime, strptime
 from .filters.url import URLFilter
 
 
 class Template(object):
+    """
+    This class represents a template for rendering content using Jinja2.
 
-    """Pickups key-value from `mkdocs.yml`"""
+    It extracts keywords from the `mkdocs.yml` configuration and provides methods to work with Jinja2 templates.
+    """
 
     __KEYS = [
         "author",
@@ -23,23 +27,36 @@ class Template(object):
         "site_url",
     ]
 
-    def __init__(self, options: object, config: Config):
+    def __init__(self, options: object, config: MkDocsConfig) -> None:
+        """
+        Initialize the Template instance.
+
+        :param options: An object containing options for the template.
+        :param config: The MkDocs configuration.
+        """
         self._options = options
         self._config = config
-
         self._keywords = None
         self._jinja_env = None
 
     @property
     def _env(self) -> jinja2.Environment:
-        def generate():
-            base_path = Path(Path(__file__).parent).resolve()
+        """
+        Get the Jinja2 environment for template rendering.
+
+        :return: The Jinja2 environment.
+        """
+
+        def generate_environment() -> jinja2.Environment:
+            """
+            Generate a Jinja2 template environment.
+
+            :return: Jinja2 template environment.
+            """
+            base_path = Path(__file__).parent.resolve()
             template_paths = []
 
-            docs_src_dir = Path(Path(self._config["config_file_path"]).parent).resolve()
-            custom_template_path = Path(self._options.custom_template_path)
-            if not custom_template_path.is_absolute():
-                custom_template_path = docs_src_dir.joinpath(self._options.custom_template_path)
+            custom_template_path = self._get_custom_template_path()
 
             if custom_template_path.is_dir():
                 template_paths.append(custom_template_path)
@@ -57,45 +74,65 @@ class Template(object):
 
             env.filters["strptime"] = strptime
             env.filters["strftime"] = strftime
-
             env.filters["to_url"] = URLFilter(self._options, self._config)
 
             return env
 
         if not self._jinja_env:
-            self._jinja_env = generate()
+            self._jinja_env = generate_environment()
         return self._jinja_env
 
     @property
-    def keywords(self) -> dict:
-        """Keywords to pass when rendering the template."""
+    def keywords(self) -> Dict:
+        """
+        Get the keywords to pass when rendering the template.
+
+        :return: A dictionary of keywords.
+        """
 
         import html
 
-        def unescape_html_in_list(values: list) -> list:
+        def unescape_html_in_list(values: List) -> List:
+            """
+            Recursively unescape HTML entities in a list of values.
+
+            :param values: The list of values.
+            :return: The list with HTML entities unescaped.
+            """
             new_values = []
             for v in values:
                 if isinstance(v, str):
                     new_values.append(html.unescape(v))
-                elif isinstance(v, list):
+                elif isinstance(v, List):
                     new_values.append(unescape_html_in_list(v))
-                elif isinstance(v, dict):
+                elif isinstance(v, Dict):
                     unescape_html(v)
                     new_values.append(v)
                 else:
                     new_values.append(v)
             return new_values
 
-        def unescape_html(variables: dict):
+        def unescape_html(variables: Dict) -> None:
+            """
+            Recursively unescape HTML entities in a dictionary of variables.
+
+            :param variables: The dictionary of variables.
+            :type variables: dict
+            """
             for k, v in variables.items():
                 if isinstance(v, str):
                     variables[k] = html.unescape(v)
-                elif isinstance(v, list):
+                elif isinstance(v, List):
                     variables[k] = unescape_html_in_list(v)
-                elif isinstance(v, dict):
+                elif isinstance(v, Dict):
                     unescape_html(v)
 
-        def build_keywords():
+        def build_keywords() -> Dict[str, Any]:
+            """
+            Build and return a dictionary of keywords.
+
+            :return: The dictionary of keywords.
+            """
             # keywords = {}
             keywords = self._config["extra"]
 
@@ -123,8 +160,17 @@ class Template(object):
 
         return self._keywords
 
-    def select(self, names: [str], parent=None, globals=None) -> jinja2.Template:
-        """Find and load a template by names of given."""
+    def select(
+        self, names: [str], parent: Optional[str] = None, globals: Optional[MutableMapping[str, Any]] = None
+    ) -> jinja2.Template:
+        """
+        Find and load a template by names.
+
+        :param names: A list of template names to search for.
+        :param parent: The parent template.
+        :param globals: Global variables to pass to the template.
+        :return: The selected Jinja2 template.
+        """
 
         real_names = []
         for name in names:
@@ -132,3 +178,15 @@ class Template(object):
                 real_names.append(name + ext)
 
         return self._env.select_template(real_names, parent=parent, globals=globals)
+
+    def _get_custom_template_path(self) -> Path:
+        """
+        Get the resolved custom template path.
+
+        :return: Resolved custom template path.
+        """
+        docs_src_dir = Path(self._config["config_file_path"]).parent.resolve()
+        custom_template_path = Path(self._options.custom_template_path)
+        if not custom_template_path.is_absolute():
+            custom_template_path = docs_src_dir.joinpath(self._options.custom_template_path)
+        return custom_template_path