+++ docs/docs/changelog.md
@@ -3,7 +3,7 @@ pdf:
   filename: Changelog
   title: Changelog & License for MkDocs PDF Generate Plugin
   type: Changelog
-  revision: 0.2.1   
+  revision: 0.2.2   
   toc_txt: true
   csv_name: Plugin Changelog
 ---
@@ -14,7 +14,7 @@ pdf:
 
 MIT License
 
-Copyright (c) 2022 Randy Duodu
+Copyright (c) 2023 Randy Duodu
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -36,6 +36,17 @@ SOFTWARE.
 
 # Changelog
 
+0.2.2
+-----
+Released: **under development**
+
+* Introduced the `cover_image` local option which allows you to add an image to the cover page of your PDF document.
+* Added code comments and docstrings to help contributors understand the code base.
+* Introduced the `include_legal_terms` global option and the `legal_terms` local option that allows the user to insert customised legal information and disclaimer into generated PDF documents'.
+* Updated documentation.
+* Fix MkDocs build failure issue.
+* Improvement: Compatibility with MkDocs >= 1.5 and Markdown 3.4.4. Fix for issue [#20](https://github.com/iSOLveIT/mkdocs-pdf-generate/issues/20).
+
 0.2.1
 -----
 Released: **22-03-2023**
@@ -45,12 +56,12 @@ Released: **22-03-2023**
 * Improvement: Added feature to generate a CSV file for all documents with the `toc_txt` local option set to `true`.
 * Improvement: Added changes that appends the revision number of a document to the filename.
 * Bugfix: Added default URL to use if `site_url` is not defined in mkdocs.yml. Fix for issue [#15](https://github.com/iSOLveIT/mkdocs-pdf-generate/issues/15).
-* Improvement: Added the `toc_txt` local option that allows users to build TXT file containing the Table of Contents
+* Improvement: Added the `toc_txt` local option that allows users to build TXT file containing the Table of Contents.
 * Minor improvements: Rewrote logging messages.
 
 0.2.0
 -----
 Released: **14-02-2023**
 
-* Package setup: Changed package from using setuptools to pyproject.toml
+* Package setup: Changed package from using setuptools to pyproject.toml.